Reward Testing
using orbit_testbed_and_visuals.jl as a testbed to play with orbits, the propagators, and to generate visuals. I've got a pretty good idea of what I want the reward function to look like now. As of the moment, sim2 and simplesim don't do much. I like Pluto for  the visuals we can do