wrote a new file initializing the basics of our MDP
So far I've defined S, R, A, and the theory for T but haven't written this.

R could use a little tweaking, the collission error is pretty good but I did just position to the ellipse as our good value function, and I think we should make it more a funciton of how close it is to keplerian elements.

We need to define a good planning strategy for approaching this problem, as well as engineering a set of satellites that will actually collide. I simualted a bunch of them and because space is big when you simulate random ones they almost never collide on their own.